Added AdvisorManager and History links to the AppBar, and gave them AppBars to link them back to the rest of the application
NOTE: This layout is not set in stone. I mainly did this for testing purposes. Eventually, Niall's idea is to have a hamburger menu in the top right, but just below the AppBar - maybe clicking on it would slide a menu from the right that runs from the bottom of the AppBar to the bottom of the screen with ElevatedButtons or something similar labelled "Course Selection", etc.